services/api/requirements.txt
django==2.1.2

djangorestframework==3.8.2

drf-yasg==1.11.0

psycopg2-binary==2.7.5

coverage==4.5.1